[MODEL] Refactored the {Searching} module and the {Response} class to allow lazy loading
Previously, the Searching::ClassMethods#search method executed the search request immediately,
and passed the Elasticsearch response as a Hash to the Response class initializer.

This prevented for instance the common pattern of paginating search results in a Rails controller action:

    @articles = Article.search(params[:q]).page(params[:page]).records

Here, the `search` method is re-initialized on each page load, immediately executing the search,
and the `#page` call then performs another, separate request.

With this patch, the Response class is initialized lazily, and the search request is performed
only when `Response#response` is called (eg. via `Response#results`, `Response#records`, etc).