[STORE] Added, that model can be created with a custom ID, and _index/_type/_version are set per model instance
Previously, it was not possible to set a custom ID for a model, when creating/saving it.
This patch fixes this ridiculous oversight.

Also, when the model is saved into a *different* index than the class-level `index_name`, the `_index`
method returns it correctly. Also applies to `_type` and `_version`.