[MODEL] Added a `HashWrapper` class to wrap Hash structures instead of raw `Hashie::Mash`
Since Hashie now prints out a warning whenever a key which clashes with a built-in method
is added, such as `sort`, the logs are flooded.

Hashie has a `disable_warnings` method in intridea/hashie@b2f94a4 to supress
the logging, so a simple sub-class has been added to allow calling it cleanly,
and to serve as a possible extension point in the future.

Many thanks to @michaelherold for the pointers to a similar implementation
in the Omniauth gem and adding quickly the infrastructure in the Hashie gem!

Closes #666